Minor updates on stochastic scripts
Correction of the comment about the variable "test" : it does not affect the number of samples.

Displacement of the assignment of the Bravais-Pearson correlation coefficient near the top of script for visibility.

Addition the variable "numWorkers" because each workers uses a lot of RAM and the computing stations are limited in RAM. Thus it is not possible to use all of the workers for now (The RAM usage of the deterministic scripts should be reduced to use more workers). Plus, it is not a good idea to use all of the workers as the computation could then easily crash.