Update phasefieldStoLinElasSingleEdgeCrack.m
Adding the display of variances and removing some remaining errors (the "'_sample_' num2str(k)").

Caution : when we compute 'epsilon_mean' from the mean of the displacement it is not equal to the mean of the epsilons.
It is the same for 'sigma_mean', the '..._von_mises_mean' and the 'internal_energy_mean'.