Add functions to test the influence of phase field model parameters
These scripts are a step to find relevant quantities of interest.
The scripts plot the load-displacement curve and compute statistics on the critical load for a range of variation of one the parameter among the Young modulus 'E', the Poisson coefficient 'NU', the fracture toughness 'gc' and the regularization length 'l'.