Update test_convergence_Shinozuka.m
For loops on the number of realizations 'N' and the one-dimensional order 'nu' have been added.
The result files are now saved in a directory containing the two parameters 'N' and the d-dimensional 'order' (order = nu^d).
The elapsed times and relative errors are written in file named 'results.txt'.