Update test_convergence_Shinozuka.m
The parameters 'N' and 'nu' are defined as vectors and the variables used in the for loops are renamed by 'Ni' and 'nui'.

The analytic autocorrelation is not computed and printed in every iteration but only at the beginning.
The order of the approximation is also printed in the printed informations.
Better management of paths and files using loop variable names 'pathnamei' and 'pathResultsi' as well as 'pathname' for the common path.
The instruction 'fclose' has been placed in the correct place.
Useless 'clf' have been got rid off.
Grids have been added in the plots.