Update test_convergence_Shinozuka and add autoCorrVec
the function 'autoCorrVec' allows to compute only the columns of interest of the autocorrelation matrix, it doesn't compute the entire covariance matrix either, allowing to compute the correlation of every point of the mesh to one point of the mesh for display purposes.
This reduces drastically the memory needed to get the wanted autocorrelation and makes the computation possible (the entire matrix needed more than 400Go of storage).

In the script 'test_convergence_Shinozuka' the previous function is now used and the displays have been improved.