Add 'myPhasefieldStoLinElasSingleEdgeCrack' and associated routine functions
The main script 'myPhasefieldStoLinElasSingleEdgeCrack' computes 'NMC' crack propagations in random material and returns statistics about the results like the mean damage field using Monte-Carlo method.
It also returns snapshots of damage field during crack propagation.
The snapshots to store are given by 'setTimeSchemeSingleEdgeCrack'.

The material properties as well as the phase field parameters can be deterministic homogeneous or random homogeneous or a combination of both or alternatively deterministic homogeneous or random heterogeneous or a combination of both.
All random properties are either homogeneous or heterogeneous, this script doesn't allow a combination of homogeneous random parameters and heterogeneous random parameters.
The level of statistical fluctuation is determined by the coefficients of variations 'deltaC' for material properties and 'deltaP1' and 'deltaP2' for phase field parameters (resp. the fracture toughness and regularization length).
The correlation lengths are stored in 'lcorr'. In order to have homogeneous random parameters, 'lcorr' must be set to '0'.